compositor: Use generated DBus adaptors from protocol XML
Makes the DBus code more maintainable and stable since the protocol is
explicitly documented in an XML file rather than inferred from class
properties and scriptable slots.

A base `DBusInterface` class is introduced which handles the initialization
so that the actual types exposed to QML (e.g. `TaskSwitcherInterface`)
are very minimalist. Since Qt DBus does not automatically handle the DBus
properties changed signal, code is added to do that manually.

The DBus API itself remains unaltered, clients continue to function
just fine.

Unfortunately, QMake does not allow specifying the parent class for
the generated adaptor (so that it is compile-time checked that all
methods are implemented), necessitating the use of boilerplate flags.