compositor: Port Notifications model to DBus XML
This simplifies the code by merging the three classes into one
with DBus outsourced to the generated DBus adaptor.

It also embraces the features of `QAbstractListModel` fully,
with the Notification data no longer exposed as an object to QML
but rather proper model roles being used. The functions on the
notification data are turned into invokables on the model itself
taking a `QModelIndex`.

The model is also made more standards-compliant:

* Actions is a list of pairs of name - label, not a flat
  list of labels
* replaces_id to replace an existing notification is supported,
  to enable using various Notification tester tools for ease
  of development

The notification UI makes use of modern QML language features, such
as `required property` instead of relying on the magic
`model` context property as well as JavaScript functions with
named arguments for signal handlers.