remove incorrect build settings
qt_build_config was introduzced at some point because it appeared to fix
building of the wayland qpa shell plugin if I remember correctly.
However it also introduces a lot of unrelated build flags, one of them
create_prl which causes all of our test clients to link against
Qt5XkbCommonSupport and Qt5ServiceSupport which is not required.