WaylandInputRegion: Disconnect ready signal once done
The QSharedPointer is copied into the lambda, thus has a ref-count of 2.
The original falls out of scope, and we're left with a ref-count of 1.

The connection is never severed, so the grab result remains alive indefinitely.

Explicitly disconnect the lambda once we're done with it, so the ref-count
drops to zero, and we properly dispose of the QQuickItemGrabResult.

(In Qt 6 we could use Qt::SingleShotConnection.)