Keyboard: Use State for keyboard animation
Merely animating the y property can cause the keyboard to be visible
when screen size changes or on application startup since the keyboard
will fly to the new position.

Instead, use an explicit State for visible and hidden and make use
of Transitions. This way the animation occurs only during a state change
from hidden to visible and vice-versa but never if the position changes
by other means.