Introduce implicit size concept for surfaces (#47)
* embedded-compositor: Add object names for areas

Makes it easier to debug what area a surface is in since
console.log will print the object name.

* Introduce implicit size concept for surfaces

Using the set_size request, a client communicates its "desired size".
This allows to distinguish between the client's requests and the
compositor's configuration. This concept is also employed by layershell
and xdg popup.

The size is mapped to a new implicitWidth and implicitHeight property
on EmbeddedShellWindow to match what other QtQuick items use for
"desired size".

Right now, width and height tries to reflect both the acutal window size
as configured by the compositor *and* the client's desired window size.
For anchored windows, however, we want the window resized by the compositor
in the anchored direction but its other axis determined by the client.

Therefore, there's a race between the compositor forcing the window to
a specific size and the client in turn seting its new desired size based
on e.g. a new screen size after rotation. Who wins is undefined.

With this, for a bottom client, you would want to set anchor to "bottom",
and implicitHeight to the desired height. Setting implicitWidth in this
scenario is ignored (unlike layershell no protocol error is raised)
since the width is enforced by the compositor to span the full width
of the bottom area.

For compatibility with existing clients, the surface's actual width or
height is used as a fallback when no respective implicit width or height
is set.

* embedded-compositor: Compress resize handling

Tries to eliminate redundant resize calls, particularly since
usually width changes and then height.