shellintegration: Rewrite to be inline with upstream
This reimplements EmbeddedShellIntegration to be in line with
how upstream QWaylandXdgShellIntegration works.

EmbeddedShellIntegration now uses QWaylandShellIntegrationTemplate
and inherits embedded_shell, too. This ensures proper initialization
of the Wayland side.

The underlying EmbeddedShell is created once the extension has been
initialized.

This makes it work under Qt 6 and removes the need for the
processEvents loop in EmbeddedShellIntegrationPlugin which in Qt 6
causes the extension to never be initialized, entering an infinite
loop and clients not starting up properly.